add slice loader pipeline
